+++ include/MeshKernel/Smoother.hpp
@@ -96,7 +96,7 @@ namespace meshkernel
 
         /// Computes operators of the elliptic smoother by node (orthonet_comp_operators)
         /// @param[in] currentNode
-        void ComputeOperatorsNode(int currentNode);
+        void ComputeOperatorsNode(size_t currentNode);
 
         /// @brief Computes m_faceNodeMappingCache, m_sharedFacesCache, m_connectedNodes for the current node, required before computing xi and eta
         /// @param[in] currentNode
@@ -110,9 +110,9 @@ namespace meshkernel
         /// @param[in] currentNode
         /// @param[in] numSharedFaces
         /// @param[in] numConnectedNodes
-        void ComputeNodeXiEta(int currentNode,
-                              int numSharedFaces,
-                              int numConnectedNodes);
+        void ComputeNodeXiEta(size_t currentNode,
+                              size_t numSharedFaces,
+                              size_t numConnectedNodes);
 
         /// @brief Compute optimal edge angle
         /// @brief numFaceNodes
@@ -133,9 +133,9 @@ namespace meshkernel
         /// @param[in] currentNode
         /// @param[in] numSharedFaces
         /// @param[in] numConnectedNodes
-        void SaveNodeTopologyIfNeeded(int currentNode,
-                                      int numSharedFaces,
-                                      int numConnectedNodes);
+        void SaveNodeTopologyIfNeeded(size_t currentNode,
+                                      size_t numSharedFaces,
+                                      size_t numConnectedNodes);
 
         /// @brief Computes local coordinates jacobian from the mapped jacobians m_Jxi and m_Jeta
         /// @param[in] currentNode
